+++ CHANGELOG.md
@@ -5,8 +5,29 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.1] - 2020-01-25
+### Added
+- Added new plotting backend using the `plotters` crate. Implementation generously provided by Hao 
+  Hou, author of the `plotters` crate.
+- Added `--plotting-backend` command-line option to select the plotting backend. The existing 
+  gnuplot backend will be used by default when available, and the plotters backend will be used when
+  gnuplot is not available or when requested.
+- Added `Criterion::plotting_backend()` function to configure the plotting backend in code.
+- Added `--load-baseline` command-line option to load a baseline for comparison
+  rather than measuring the current code
+- Benchmark filters can now be regular expressions.
+
 ### Fixed
+- Fixed `fibonacci` functions.
 - Fixed `#[criterion]` benchmarks ignoring the command-line options.
+- Fixed incorrect scaling of the violin plots.
+- Don't print the recommended sample count if it's the same as the configured
+  sample count.
+- Fix potential panic when `nresamples` is set too low. Also added a warning
+  against setting `nresamples` too low.
+- Fixed issue where a slow outer closure would cause Criterion.rs to calculate 
+  the wrong estimated time and number of iterations in the warm-up phase.
 
 ## [0.3.0] - 2019-08-25
 ### Added
@@ -293,7 +314,7 @@ more details
 - Initial release on Crates.io.
 
 
-[Unreleased]: https://github.com/bheisler/criterion.rs/compare/0.3.0...HEAD
+[Unreleased]: https://github.com/bheisler/criterion.rs/compare/0.3.1...HEAD
 [0.1.1]: https://github.com/bheisler/criterion.rs/compare/0.1.0...0.1.1
 [0.1.2]: https://github.com/bheisler/criterion.rs/compare/0.1.1...0.1.2
 [0.2.0]: https://github.com/bheisler/criterion.rs/compare/0.1.2...0.2.0
@@ -309,3 +330,4 @@ more details
 [0.2.10]: https://github.com/bheisler/criterion.rs/compare/0.2.9...0.2.10
 [0.2.11]: https://github.com/bheisler/criterion.rs/compare/0.2.10...0.2.11
 [0.3.0]: https://github.com/bheisler/criterion.rs/compare/0.2.11...0.3.0
+[0.3.1]: https://github.com/bheisler/criterion.rs/compare/0.3.0...0.3.1